Fix error message in Font scale plugin when system settings write management is not available
Add readme sections for FoQA plugins
