+++ README.md
@@ -1,6 +1,6 @@
 # FoQA
 
-<img src="seal.svg" alt="Logo" width="200"/>
+<img src="art/seal.svg" alt="Logo" width="200"/>
 
 Container for various Quality Assurance utilities to be included in QA/testing variants of Android apps.
 
@@ -31,8 +31,22 @@ repositories {
 - [Hyperion Chuck](https://github.com/Commit451/Hyperion-Chuck)
 - [Hyperion App Info](https://github.com/willowtreeapps/Hyperion-Android)
 - [Android Debug Database](https://github.com/amitshekhariitbhu/Android-Debug-Database)
-- Hyperion menu item with Device market name [DeviceInfoPlugin](device_info_plugin)
-- Hyperion menu item with font scale changer [FontScalePlugin](font_scale_plugin)
+- Hyperion menu item with Device market name [Hyperion Device Info](device_info_plugin)
+- Hyperion menu item with font scale changer [Hyperion Font Scale](font_scale_plugin)
+
+### Device Info plugin
+
+Shows device market name and Android version.
+
+<img src="art/device_info_pixel.png" alt="Device info plugin demo"/>
+<img src="art/device_info_sgs4.png" alt="Device info plugin demo"/>
+
+### Font Scale plugin
+
+Allows to quickly change system font scale in range from `0.4` to `1.6`.
+On devices with Android 6 or newer system settings write prompt will appear when setting scale for the first time. 
+
+<img src="art/font_scale.gif" alt="Font scale plugin demo"/>
 
 See [build.gradle](library/build.gradle) for a complete list of components. And [build.gradle](build.gradle) for versions.
