+++ font_scale_plugin/src/main/java/pl/droidsonroids/foqa/deviceinfo/FontScaleModule.kt
@@ -1,17 +1,10 @@
 package pl.droidsonroids.foqa.deviceinfo
 
-import android.content.ActivityNotFoundException
-import android.content.Context
-import android.content.Intent
-import android.os.Build
-import android.provider.Settings
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.SeekBar
 import android.widget.TextView
-import android.widget.Toast
-import androidx.annotation.RequiresApi
 import com.willowtreeapps.hyperion.plugin.v1.PluginModule
 import pl.droidsonroids.foqa.fontscale.R
 
@@ -66,33 +59,4 @@ internal class FontScaleModule : PluginModule() {
     }
 
     override fun getName() = R.string.foqa_font_scale_plugin_name
-}
-
-internal class FontScaleInteractor(private val context: Context) {
-
-    var systemFontScale
-        get() = Settings.System.getFloat(context.contentResolver, Settings.System.FONT_SCALE)
-        set(value) {
-            Settings.System.putFloat(context.contentResolver, Settings.System.FONT_SCALE, value)
-        }
-
-    val isSystemSettingsWritable
-        get() = Build.VERSION.SDK_INT < Build.VERSION_CODES.M || Settings.System.canWrite(context)
-
-    fun ensureSystemSettingsWritability() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M && !Settings.System.canWrite(context)) {
-            startSystemSettingsActivity()
-        }
-    }
-
-    @RequiresApi(Build.VERSION_CODES.M)
-    private fun startSystemSettingsActivity() {
-        val intent = Intent(Settings.ACTION_MANAGE_WRITE_SETTINGS)
-            .addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-        try {
-            context.startActivity(intent)
-        } catch (e: ActivityNotFoundException) {
-            Toast.makeText(context, R.string.foqa_font_scale_plugin_name, Toast.LENGTH_SHORT).show()
-        }
-    }
 }
