+++ font_scale_plugin/src/main/res/values/values.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="foqa_font_scale_plugin_name">Font Scale</string>
+    <string name="foqa_font_scale_system_settings_disabled">System settings write management not available</string>
     <string name="foqa_font_scale_value">%1.2f</string>
 </resources>
